server: send request to dest
