+++ src/client/src/node_client.rs
@@ -75,6 +75,12 @@ impl Client {
         let res = client.pull(req).await?;
         Ok(res.into_inner())
     }
+
+    pub async fn forward(&self, req: ForwardRequest) -> Result<ForwardResponse, tonic::Status> {
+        let mut client = self.client.clone();
+        let res = client.forward(req).await?;
+        Ok(res.into_inner())
+    }
 }
 
 #[derive(Debug, Clone)]
