+++ src/server/src/node/replica/eval/cmd_delete.rs
@@ -16,7 +16,8 @@ use engula_api::server::v1::ShardDeleteRequest;
 use crate::{
     node::{
         engine::{GroupEngine, WriteBatch},
-        replica::ExecCtx, migrate::ForwardCtx,
+        migrate::ForwardCtx,
+        replica::ExecCtx,
     },
     serverpb::v1::{EvalResult, WriteBatchRep},
     Error, Result,
