+++ src/server/src/node/replica/eval/cmd_put.rs
@@ -17,7 +17,8 @@ use engula_api::server::v1::ShardPutRequest;
 use crate::{
     node::{
         engine::{GroupEngine, WriteBatch},
-        replica::ExecCtx, migrate::ForwardCtx,
+        migrate::ForwardCtx,
+        replica::ExecCtx,
     },
     serverpb::v1::{EvalResult, WriteBatchRep},
     Error, Result,
